+++ packages/components/src/Form/Inputs/InputTime/InputTime.tsx
@@ -41,13 +41,13 @@ import add from 'lodash/add'
 import subtract from 'lodash/subtract'
 
 import {
-  InputText,
   inputTextHover,
   inputTextFocus,
   inputTextDisabled,
   inputTextValidation,
   inputCSS,
 } from '../InputText'
+import { innerInputStyle } from '../innerInputStyle'
 import { Icon } from '../../../Icon'
 import {
   simpleLayoutCSS,
@@ -469,10 +469,8 @@ const InputTimeInternal = forwardRef(
         onBlur={onBlur}
         aria-invalid={validationType === 'error' ? 'true' : undefined}
       >
-        <InputText
+        <StyledInput
           id={id}
-          maxLength={2}
-          placeholder="--"
           value={hour}
           onKeyDown={readOnly ? noop : handleHourKeyDown}
           onFocus={handleHourFocus}
@@ -485,9 +483,7 @@ const InputTimeInternal = forwardRef(
           required={required}
         />
         <div>:</div>
-        <InputText
-          maxLength={2}
-          placeholder="--"
+        <StyledInput
           value={minute}
           onKeyDown={readOnly ? noop : handleMinuteKeyDown}
           onFocus={handleMinuteFocus}
@@ -500,9 +496,7 @@ const InputTimeInternal = forwardRef(
           required={required}
         />
         {format === '12h' ? (
-          <InputText
-            maxLength={2}
-            placeholder="--"
+          <StyledInput
             value={period}
             onKeyDown={readOnly ? noop : handlePeriodKeyDown}
             onFocus={handlePeriodFocus}
@@ -532,6 +526,24 @@ const InputTimeInternal = forwardRef(
 
 const WarningIcon = styled(Icon)``
 
+const StyledInput = styled.input.attrs({
+  maxLength: 2,
+  placeholder: '--',
+  type: 'text',
+})`
+  ${innerInputStyle}
+  font-size: ${(props) => props.theme.fontSizes.small};
+  height: 34px;
+  line-height: ${({ theme }) => theme.lineHeights.medium};
+  padding: 0;
+  text-align: center;
+  width: 1.25rem;
+
+  &:focus {
+    background: ${({ theme }) => theme.colors.keyAccent};
+  }
+`
+
 export const InputTime = styled(InputTimeInternal)`
   ${simpleLayoutCSS}
   ${inputCSS}
@@ -541,27 +553,6 @@ export const InputTime = styled(InputTimeInternal)`
   grid-gap: 0.15rem;
   grid-template-columns: auto auto auto auto 1fr;
   padding: 0 ${({ theme }) => theme.space.xsmall};
-
-  ${InputText} {
-    background: transparent;
-    border: none;
-    border-radius: 0;
-    box-shadow: none;
-    color: inherit;
-    height: 34px;
-    line-height: ${({ theme }) => theme.lineHeights.medium};
-    margin: 0;
-    padding: 0;
-    width: 1.25rem;
-
-    input {
-      text-align: center;
-    }
-
-    &:focus-within {
-      background: ${({ theme }) => theme.colors.keyAccent};
-    }
-  }
 
   ${WarningIcon} {
     justify-self: end;